Totalmente funcional la lista de compras

+++ AppIncalink/Models/recetasModel.cs
@@ -5,7 +5,7 @@
         public int id { get; set; }
         public int idMenu { get; set; }
         public int idProducto { get; set; }
-        public float cantidad { get; set; }
+        public decimal cantidad { get; set; }
     
     }
